+++ Runtime/Registry/MassiveRegistry.cs
@@ -6,16 +6,14 @@ namespace Massive
 	public class MassiveRegistry : Registry, IMassive
 	{
 		private readonly MassiveEntities _massiveEntities;
-		private readonly MassiveGroupRegistry _massiveGroups;
 
 		public MassiveRegistry(int setCapacity = Constants.DefaultSetCapacity, int framesCapacity = Constants.DefaultFramesCapacity,
 			bool storeEmptyTypesAsDataSets = false, int pageSize = Constants.DefaultPageSize)
-			: base(new MassiveGroupRegistry(setCapacity, framesCapacity), new MassiveEntities(setCapacity, framesCapacity),
-				new MassiveSetFactory(setCapacity, framesCapacity, storeEmptyTypesAsDataSets, pageSize))
+			: base(new MassiveSetFactory(setCapacity, framesCapacity, storeEmptyTypesAsDataSets, pageSize),
+				new MassiveGroupFactory(setCapacity, framesCapacity), new MassiveEntities(setCapacity, framesCapacity))
 		{
 			// Fetch instances from base
 			_massiveEntities = (MassiveEntities)Entities;
-			_massiveGroups = (MassiveGroupRegistry)GroupRegistry;
 		}
 
 		public int CanRollbackFrames => _massiveEntities.CanRollbackFrames;
@@ -24,11 +22,18 @@ namespace Massive
 		public void SaveFrame()
 		{
 			_massiveEntities.SaveFrame();
-			_massiveGroups.SaveFrame();
 
 			for (var i = 0; i < SetRegistry.All.Count; i++)
 			{
 				if (SetRegistry.All[i] is IMassive massive)
+				{
+					massive.SaveFrame();
+				}
+			}
+
+			for (var i = 0; i < GroupRegistry.All.Count; i++)
+			{
+				if (GroupRegistry.All[i] is IMassive massive)
 				{
 					massive.SaveFrame();
 				}
@@ -39,11 +44,18 @@ namespace Massive
 		public void Rollback(int frames)
 		{
 			_massiveEntities.Rollback(Math.Min(frames, _massiveEntities.CanRollbackFrames));
-			_massiveGroups.Rollback(Math.Min(frames, _massiveGroups.CanRollbackFrames));
 
 			for (var i = 0; i < SetRegistry.All.Count; i++)
 			{
 				if (SetRegistry.All[i] is IMassive massive)
+				{
+					massive.Rollback(Math.Min(frames, massive.CanRollbackFrames));
+				}
+			}
+
+			for (var i = 0; i < GroupRegistry.All.Count; i++)
+			{
+				if (GroupRegistry.All[i] is IMassive massive)
 				{
 					massive.Rollback(Math.Min(frames, massive.CanRollbackFrames));
 				}